+++ tests/test_pipeline_files.py
@@ -19,7 +19,6 @@ from unittest.mock import MagicMock
 import attr
 import pytest
 from itemadapter import ItemAdapter
-from testfixtures import LogCapture
 from twisted.internet.defer import Deferred
 from twisted.python.failure import Failure
 
@@ -33,9 +32,8 @@ from scrapy.pipelines.files import (
     FTPFilesStore,
     GCSFilesStore,
     S3FilesStore,
-    _MediaRequestFiltered,
 )
-from scrapy.pipelines.media import MediaPipeline
+from scrapy.pipelines.media import MediaPipeline, _MediaRequestFiltered
 from scrapy.settings import Settings
 from scrapy.utils.asyncio import call_later
 from scrapy.utils.defer import maybe_deferred_to_future
@@ -295,7 +293,9 @@ class TestFilesPipeline:
         request = Request("http://example.com")
         assert file_path(request, item=item) == "full/path-to-store-file"
 
-    def test_media_failed_filtered_request(self):
+    def test_media_failed_filtered_request(
+        self, caplog: pytest.LogCaptureFixture
+    ) -> None:
         """A filtered media request (IgnoreRequest) is reported as a
         _MediaRequestFiltered exception and logged at the DEBUG level, instead
         of as a download error with a traceback."""
@@ -304,31 +304,35 @@ class TestFilesPipeline:
         failure = Failure(IgnoreRequest(reason))
 
         with (
-            LogCapture() as log,
+            caplog.at_level(logging.DEBUG),
             pytest.raises(_MediaRequestFiltered, match=re.escape(reason)),
         ):
             self.pipeline.media_failed(failure, request, self.pipeline.spiderinfo)
 
-        assert len(log.records) == 1
-        record = log.records[0]
+        assert len(caplog.records) == 1
+        record = caplog.records[0]
         assert record.levelname == "DEBUG"
         assert record.exc_info is None
         assert reason in record.getMessage()
 
-    def test_media_failed_download_error(self):
+    def test_media_failed_download_error(
+        self, caplog: pytest.LogCaptureFixture
+    ) -> None:
         """A genuine download error is reported as a FileException and logged as
         a warning."""
         request = Request("http://example.com/file.pdf")
         failure = Failure(Exception("boom"))
 
-        with LogCapture() as log, pytest.raises(FileException):
+        with caplog.at_level(logging.WARNING), pytest.raises(FileException):
             self.pipeline.media_failed(failure, request, self.pipeline.spiderinfo)
 
-        assert len(log.records) == 1
-        assert log.records[0].levelname == "WARNING"
+        assert len(caplog.records) == 1
+        assert caplog.records[0].levelname == "WARNING"
 
     @coroutine_test
-    async def test_process_item_filtered_request(self):
+    async def test_process_item_filtered_request(
+        self, caplog: pytest.LogCaptureFixture
+    ) -> None:
         """A filtered (e.g. offsite) media request is processed as a failed
         result without being logged as an error with a traceback."""
         item_url = "http://example.com/file.pdf"
@@ -340,7 +344,7 @@ class TestFilesPipeline:
             },
         )
         with (
-            LogCapture() as log,
+            caplog.at_level(logging.DEBUG),
             mock.patch.object(
                 FilesPipeline, "get_media_requests", return_value=[request]
             ),
@@ -348,10 +352,10 @@ class TestFilesPipeline:
             result = await self.pipeline.process_item(item)
 
         assert result["files"] == []
-        assert not any(r.levelname in ("WARNING", "ERROR") for r in log.records)
+        assert not any(r.levelname in ("WARNING", "ERROR") for r in caplog.records)
         assert any(
             "Filtered offsite request to 'example.com'" in r.getMessage()
-            for r in log.records
+            for r in caplog.records
         )
 
     @pytest.mark.parametrize(
